Rename NeuralAddVectorMultipliedByValue to SynetAddVectorMultipliedByValue.
Move architecture implementations into dedicated SynetMath.cpp files, relocate
the auto-test to TestSynetMath.cpp, update VS2022 projects, and refresh release
7.2.165 notes in docs/2026.html.
