+++ docs/2026.html
@@ -54,10 +54,21 @@
 <ul>
  <li>Rename function SimdVectorProduct to SimdCreateMask.</li>
  <li>Rename function Simd::VectorProduct to Simd::CreateMask.</li>
+ <li>Rename function SimdNeuralAddVectorMultipliedByValue to SimdSynetAddVectorMultipliedByValue.</li>
+ <li>Rename function Simd::NeuralAddVectorMultipliedByValue to Simd::SynetAddVectorMultipliedByValue.</li>
+</ul>
+<h5>Improving</h5>
+<ul>
+ <li>Move Base, SSE4.1, AVX2, AVX-512BW, NEON, SVE2 optimizations of function SynetAddVectorMultipliedByValue to separate files.</li>
 </ul>
 <h5>Removing</h5>
 <ul>
  <li>Separate support of SVE extension (ARM/ARM64 platform).</li>
+</ul>
+<h4>Test framework</h4>
+<h5>Improving</h5>
+<ul>
+ <li>Move tests of function SynetAddVectorMultipliedByValue from TestNeural.cpp to TestSynetMath.cpp.</li>
 </ul>
 
 <a href="#HOME">Home</a>
@@ -690,7 +701,7 @@
  <li>Description of function SimdNeuralDerivativeRelu.</li>
  <li>Description of function SimdNeuralPow.</li>
  <li>Description of function SimdNeuralProductSum.</li>
- <li>Description of function SimdNeuralAddVectorMultipliedByValue.</li>
+ <li>Description of function SimdSynetAddVectorMultipliedByValue.</li>
  <li>Description of function SimdNeuralAddVector.</li>
  <li>Description of function SimdNeuralAddValue.</li>
  <li>Description of function SimdNeuralUpdateWeights.</li>
