+++ prj/vs2022/Avx2.vcxproj
@@ -94,6 +94,7 @@
     <ClCompile Include="..\..\src\Simd\SimdAvx2Synet.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdAvx2SynetActivation.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdAvx2SynetAdd.cpp" />
+    <ClCompile Include="..\..\src\Simd\SimdAvx2SynetMath.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdAvx2SynetAdd16b.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdAvx2SynetConversion.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdAvx2SynetConvolution16b.cpp" />
