+++ prj/vs2022/Avx512bw.vcxproj
@@ -94,6 +94,7 @@
     <ClCompile Include="..\..\src\Simd\SimdAvx512bwSynet.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdAvx512bwSynetActivation.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdAvx512bwSynetAdd.cpp" />
+    <ClCompile Include="..\..\src\Simd\SimdAvx512bwSynetMath.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdAvx512bwSynetAdd16b.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdAvx512bwSynetConversion.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdAvx512bwSynetConvolution16b.cpp" />
