+++ prj/vs2022/Base.vcxproj
@@ -218,6 +218,7 @@
     <ClCompile Include="..\..\src\Simd\SimdBaseSynet.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdBaseSynetActivation.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdBaseSynetAdd.cpp" />
+    <ClCompile Include="..\..\src\Simd\SimdBaseSynetMath.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdBaseSynetAdd16b.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdBaseSynetConversion.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdBaseSynetConvolution16b.cpp" />
