+++ prj/vs2022/Neon.vcxproj
@@ -104,6 +104,7 @@
     <ClCompile Include="..\..\src\Simd\SimdNeonSynet.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdNeonSynetActivation.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdNeonSynetAdd.cpp" />
+    <ClCompile Include="..\..\src\Simd\SimdNeonSynetMath.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdNeonSynetAdd16b.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdNeonSynetNormalize16b.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdNeonSynetScale16b.cpp" />
