+++ prj/vs2022/Test.vcxproj
@@ -114,6 +114,7 @@
     <ClCompile Include="..\..\src\Test\TestSynet.cpp" />
     <ClCompile Include="..\..\src\Test\TestSynetActivation.cpp" />
     <ClCompile Include="..\..\src\Test\TestSynetAdd.cpp" />
+    <ClCompile Include="..\..\src\Test\TestSynetMath.cpp" />
     <ClCompile Include="..\..\src\Test\TestSynetConversion.cpp" />
     <ClCompile Include="..\..\src\Test\TestSynetConvolution16b.cpp" />
     <ClCompile Include="..\..\src\Test\TestSynetConvolution32f.cpp" />
