+++ src/Simd/SimdBaseSynetMath.cpp
@@ -1,0 +1,51 @@
+/*
+* Simd Library (http://ermig1979.github.io/Simd).
+*
+* Copyright (c) 2011-2026 Yermalayeu Ihar.
+*
+* Permission is hereby granted, free of charge, to any person obtaining a copy
+* of this software and associated documentation files (the "Software"), to deal
+* in the Software without restriction, including without limitation the rights
+* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+* copies of the Software, and to permit persons to whom the Software is
+* furnished to do so, subject to the following conditions:
+*
+* The above copyright notice and this permission notice shall be included in
+* all copies or substantial portions of the Software.
+*
+* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+* FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+* AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+* LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+* OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+* SOFTWARE.
+*/
+#include "Simd/SimdMemory.h"
+
+namespace Simd
+{
+#if defined(SIMD_SYNET_ENABLE)
+    namespace Base
+    {
+        SIMD_INLINE void AddMultiplied(const float * src, size_t aligned, size_t full, float value, float * dst)
+        {
+            size_t i = 0;
+            for (; i < aligned; i += 4)
+            {
+                dst[i + 0] += src[i + 0] * value;
+                dst[i + 1] += src[i + 1] * value;
+                dst[i + 2] += src[i + 2] * value;
+                dst[i + 3] += src[i + 3] * value;
+            }
+            for (; i < full; ++i)
+                dst[i] += src[i] * value;
+        }
+
+        void SynetAddVectorMultipliedByValue(const float * src, size_t size, const float * value, float * dst)
+        {
+            AddMultiplied(src, Simd::AlignLo(size, 4), size, *value, dst);
+        }
+    }
+#endif
+}
