+++ src/Simd/SimdLib.h
@@ -5482,28 +5482,6 @@ extern "C"
 
     /*! @ingroup neural
 
-        \fn void SimdNeuralAddVectorMultipliedByValue(const float * src, size_t size, const float * value, float * dst);
-
-        \short Adds a source vector multiplied by a scalar to a destination vector.
-
-        \deprecated This function will be removed in the nearest future.
-
-        For every element:
-        \verbatim
-        dst[i] += src[i]*value[0];
-        \endverbatim
-
-        \note This function is used in Simd::Neural.
-
-        \param [in] src - a pointer to the input 32-bit float array.
-        \param [in] size - a size of arrays.
-        \param [in] value - a pointer to the scalar 32-bit float value.
-        \param [in, out] dst - a pointer to cumulative 32-bit float array.
-    */
-    SIMD_DEPRECATED SIMD_API void SimdNeuralAddVectorMultipliedByValue(const float * src, size_t size, const float * value, float * dst);
-
-    /*! @ingroup neural
-
         \fn void SimdNeuralAddVector(const float * src, size_t size, float * dst);
 
         \short Adds a source vector to a destination vector.
@@ -7611,6 +7589,26 @@ extern "C"
         \param [in] format - a format of the tensor.
     */
     SIMD_API void SimdSynetAddBias(const float * bias, size_t channels, size_t spatial, float * dst, SimdTensorFormatType format);
+
+    /*! @ingroup synet_other
+
+        \fn void SimdSynetAddVectorMultipliedByValue(const float * src, size_t size, const float * value, float * dst);
+
+        \short Adds a source vector multiplied by a scalar to a destination vector.
+
+        For every element:
+        \verbatim
+        dst[i] += src[i]*value[0];
+        \endverbatim
+
+        \note This function is used in <a href="http://github.com/ermig1979/Synet">Synet Framework</a>.
+
+        \param [in] src - a pointer to the input 32-bit float array.
+        \param [in] size - a size of arrays.
+        \param [in] value - a pointer to the scalar 32-bit float value.
+        \param [in, out] dst - a pointer to cumulative 32-bit float array.
+    */
+    SIMD_API void SimdSynetAddVectorMultipliedByValue(const float * src, size_t size, const float * value, float * dst);
 
     /*! @ingroup synet_add
