+++ src/Simd/SimdSse41Neural.cpp
@@ -89,18 +89,6 @@ namespace Simd
 
         //-------------------------------------------------------------------------------------------------
 
-        void NeuralAddVectorMultipliedByValue(const float* src, size_t size, const float* value, float* dst)
-        {
-            size_t aligned = AlignLo(size, QF);
-            size_t partial = AlignLo(size, F);
-            if (Aligned(src) && Aligned(dst))
-                AddMultiplied<true>(src, aligned, partial, size, *value, dst);
-            else
-                AddMultiplied<false>(src, aligned, partial, size, *value, dst);
-        }
-
-        //-------------------------------------------------------------------------------------------------
-
         template <bool align> SIMD_INLINE void AddVector(const float* src, float* dst)
         {
             Store<align>(dst, _mm_add_ps(Load<align>(dst), Load<align>(src)));
