+++ src/Simd/SimdSve2Neural.cpp
@@ -118,30 +118,6 @@ namespace Simd
 
         //-------------------------------------------------------------------------------------------------
 
-        SIMD_INLINE void AddMultiplied(const svbool_t& mask, const svfloat32_t& value, const float* src, float* dst)
-        {
-            svst1_f32(mask, dst, svmla_f32_m(mask, svld1_f32(mask, dst), svld1_f32(mask, src), value));
-        }
-
-        void NeuralAddVectorMultipliedByValue(const float* src, size_t size, const float* value, float* dst)
-        {
-            size_t F = svcntw(), QF = 4 * F, i = 0;
-            const svbool_t body = svptrue_b32();
-            const svfloat32_t _value = svdup_n_f32(value[0]);
-
-            for (; i + QF <= size; i += QF)
-            {
-                AddMultiplied(body, _value, src + i + 0 * F, dst + i + 0 * F);
-                AddMultiplied(body, _value, src + i + 1 * F, dst + i + 1 * F);
-                AddMultiplied(body, _value, src + i + 2 * F, dst + i + 2 * F);
-                AddMultiplied(body, _value, src + i + 3 * F, dst + i + 3 * F);
-            }
-            for (; i + F <= size; i += F)
-                AddMultiplied(body, _value, src + i, dst + i);
-            if (i < size)
-                AddMultiplied(svwhilelt_b32(i, size), _value, src + i, dst + i);
-        }
-
         SIMD_INLINE void AddVector(const svbool_t& mask, const float* src, float* dst)
         {
             svst1_f32(mask, dst, svadd_f32_x(mask, svld1_f32(mask, dst), svld1_f32(mask, src)));
