+++ src/Test/TestDescrInt.cpp
@@ -787,9 +787,9 @@ namespace Test
     static inline void SetRandomDescriptor(const float* rnd, size_t size, float mainRange, int seed, float noiseRange, size_t noiseTimes, float* dst)
     {
         memset(dst, 0, size * sizeof(float));
-        SimdNeuralAddVectorMultipliedByValue(rnd + seed, size, &mainRange, dst);
+        SimdSynetAddVectorMultipliedByValue(rnd + seed, size, &mainRange, dst);
         for(size_t t = 0; t < noiseTimes; ++t)
-            SimdNeuralAddVectorMultipliedByValue(rnd + (Rand() & INT16_MAX), size, &noiseRange, dst);
+            SimdSynetAddVectorMultipliedByValue(rnd + (Rand() & INT16_MAX), size, &noiseRange, dst);
     }
 
     typedef std::pair<size_t, float> Pair;
