+++ src/Test/TestNeural.cpp
@@ -241,99 +241,6 @@ namespace Test
 
     namespace
     {
-        struct FuncAVMV
-        {
-            typedef void(*FuncPtr)(const float * src, size_t size, const float * value, float * dst);
-
-            FuncPtr func;
-            String description;
-
-            FuncAVMV(const FuncPtr & f, const String & d) : func(f), description(d) {}
-
-            void Call(const View & src, float value, const View & dstSrc, View & dstDst) const
-            {
-                Simd::Copy(dstSrc, dstDst);
-                TEST_PERFORMANCE_TEST(description);
-                func((float*)src.data, src.width, &value, (float*)dstDst.data);
-            }
-        };
-    }
-#define FUNC_AVMV(function) FuncAVMV(function, #function)
-
-    bool NeuralAddVectorMultipliedByValueAutoTest(int size, float eps, const FuncAVMV & f1, const FuncAVMV & f2)
-    {
-        bool result = true;
-
-        TEST_LOG_SS(Info, "Test " << f1.description << " & " << f2.description << " [" << size << "].");
-
-        View src(size, 1, View::Float, NULL, TEST_ALIGN(size));
-        FillRandom32f(src);
-
-        View dstSrc(size, 1, View::Float, NULL, TEST_ALIGN(size));
-        FillRandom32f(dstSrc);
-
-        const float value = 0.3f;
-
-        View dstDst1(size, 1, View::Float, NULL, TEST_ALIGN(size));
-        View dstDst2(size, 1, View::Float, NULL, TEST_ALIGN(size));
-
-        TEST_EXECUTE_AT_LEAST_MIN_TIME(f1.Call(src, value, dstSrc, dstDst1));
-
-        TEST_EXECUTE_AT_LEAST_MIN_TIME(f2.Call(src, value, dstSrc, dstDst2));
-
-        result = Compare(dstDst1, dstDst2, eps, true);
-
-        return result;
-    }
-
-    bool NeuralAddVectorMultipliedByValueAutoTest(float eps, const FuncAVMV & f1, const FuncAVMV & f2)
-    {
-        bool result = true;
-
-        result = result && NeuralAddVectorMultipliedByValueAutoTest(W*H, eps, f1, f2);
-        result = result && NeuralAddVectorMultipliedByValueAutoTest(W*H + O, eps, f1, f2);
-        result = result && NeuralAddVectorMultipliedByValueAutoTest(W*H - O, eps, f1, f2);
-
-        return result;
-    }
-
-    bool NeuralAddVectorMultipliedByValueAutoTest(const Options & options)
-    {
-        bool result = true;
-
-        if (TestBase(options))
-            result = result && NeuralAddVectorMultipliedByValueAutoTest(EPS, FUNC_AVMV(Simd::Base::NeuralAddVectorMultipliedByValue), FUNC_AVMV(SimdNeuralAddVectorMultipliedByValue));
-
-#ifdef SIMD_SSE41_ENABLE
-        if (Simd::Sse41::Enable && TestSse41(options))
-            result = result && NeuralAddVectorMultipliedByValueAutoTest(EPS, FUNC_AVMV(Simd::Sse41::NeuralAddVectorMultipliedByValue), FUNC_AVMV(SimdNeuralAddVectorMultipliedByValue));
-#endif 
-
-#ifdef SIMD_AVX2_ENABLE
-        if (Simd::Avx2::Enable && TestAvx2(options))
-            result = result && NeuralAddVectorMultipliedByValueAutoTest(EPS, FUNC_AVMV(Simd::Avx2::NeuralAddVectorMultipliedByValue), FUNC_AVMV(SimdNeuralAddVectorMultipliedByValue));
-#endif
-
-#ifdef SIMD_AVX512BW_ENABLE
-        if (Simd::Avx512bw::Enable && TestAvx512bw(options))
-            result = result && NeuralAddVectorMultipliedByValueAutoTest(EPS, FUNC_AVMV(Simd::Avx512bw::NeuralAddVectorMultipliedByValue), FUNC_AVMV(SimdNeuralAddVectorMultipliedByValue));
-#endif
-
-#ifdef SIMD_SVE2_ENABLE
-        if (Simd::Sve2::Enable && TestSve2(options))
-            result = result && NeuralAddVectorMultipliedByValueAutoTest(EPS, FUNC_AVMV(Simd::Sve2::NeuralAddVectorMultipliedByValue), FUNC_AVMV(SimdNeuralAddVectorMultipliedByValue));
-#endif
-
-#ifdef SIMD_NEON_ENABLE
-        if (Simd::Neon::Enable && TestNeon(options))
-            result = result && NeuralAddVectorMultipliedByValueAutoTest(EPS, FUNC_AVMV(Simd::Neon::NeuralAddVectorMultipliedByValue), FUNC_AVMV(SimdNeuralAddVectorMultipliedByValue));
-#endif
-
-        return result;
-    }
-
-    namespace
-    {
         struct FuncAddVec
         {
             typedef void(*FuncPtr)(const float * src, size_t size, float * dst);
